fix(Cashier): fix clearing SalesCard list, sync with store
Weird interaction. If we assign `cart.cart` value to a variable like
```
const currentSale: CartItem[] = cart.cart
```
`currentSale` value is not syncing anymore with the store if we clear it.
It sync just fine if we push item to `cart.cart` and `currentSale` will be
updated accordingly. But if we clear `cart.cart` value through `cart.$reset`
or `cart.cart = []` or even `cart.$patch({})`, `currentSale` will lose it's
*reactivity*.

Is it a bug? or feature?

Some issues I came across while debugging this

https://github.com/vuejs/pinia/issues/593
https://github.com/vuejs/pinia/issues/484

After a bit more research, found this.

https://github.com/vuejs/pinia/discussions/1191
https://pinia.vuejs.org/core-concepts/#using-the-store

Haven't test wrapping it with `computed` or `storeToRefs` yet